+++ src/main/java/com/ems/from/BusinessSearchForm.java
@@ -3,6 +3,7 @@ package com.ems.from;
 import org.springframework.format.annotation.DateTimeFormat;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Date;
 
 public class BusinessSearchForm extends PageSearchForm implements Serializable{
@@ -14,7 +15,7 @@ public class BusinessSearchForm extends PageSearchForm implements Serializable{
 
     private Long userId;
 
-    private Integer fraction;
+    private BigDecimal fraction;
 
     private Integer businessCount;
 
@@ -52,11 +53,11 @@ public class BusinessSearchForm extends PageSearchForm implements Serializable{
         this.userId = userId;
     }
 
-    public Integer getFraction() {
+    public BigDecimal getFraction() {
         return fraction;
     }
 
-    public void setFraction(Integer fraction) {
+    public void setFraction(BigDecimal fraction) {
         this.fraction = fraction;
     }
